+++ modules/core/src/lib/common.ts
@@ -123,15 +123,14 @@ export function degrees(
   return map(radians, (radians) => radians * RADIANS_TO_DEGREES, result);
 }
 
-
 /**
  * The modulo operation that also works for negative dividends.
  *
- * @param {number} m The dividend.
- * @param {number} n The divisor.
- * @returns {number} The remainder.
+ * @param m The dividend.
+ * @param n The divisor.
+ * @returns The remainder.
  */
-export function mod(m: number, n: number): number {
+export function safeMod(m: number, n: number): number {
   if (Math.sign(m) === Math.sign(n) && Math.abs(m) < Math.abs(n)) {
     return m;
   }
@@ -140,20 +139,38 @@ export function mod(m: number, n: number): number {
 }
 
 /**
+ * Produces an angle restricted to its equivalent in a normalized range
+ *
+ * @param angle in radians
+ * @param range 'zero-to-two-pi' - in the range 0 <= angle <= 2PI, | 'negative-pi-to-pi' -  -Pi <= angle <= Pi
+ * @returns The angle in the range [0, <code>TWO_PI</code>] or  [<code>-PI</code>, <code>PI</code>]..
+ */
+export function normalizeAngle(
+  angle: number,
+  range: 'zero-to-two-pi' | 'negative-pi-to-pi'
+): number {
+  switch (range) {
+    case 'negative-pi-to-pi':
+      return negativePiToPi(angle);
+    case 'zero-to-two-pi':
+      return zeroToTwoPi(angle);
+    default:
+      return angle;
+  }
+}
+
+/**
  * Produces an angle in the range 0 <= angle <= 2Pi which is equivalent to the provided angle.
  *
- * @param {number} angle in radians
- * @returns {number} The angle in the range [0, <code>TWO_PI</code>].
+ * @param angle in radians
+ * @returns The angle in the range [0, <code>TWO_PI</code>].
  */
-export function zeroToTwoPi(angle: number): number {
+function zeroToTwoPi(angle: number): number {
   if (angle >= 0 && angle <= TWO_PI) {
     return angle;
   }
-  const remainder = mod(angle, TWO_PI);
-  if (
-      Math.abs(remainder) < EPSILON14 &&
-      Math.abs(angle) > EPSILON14
-  ) {
+  const remainder = safeMod(angle, TWO_PI);
+  if (Math.abs(remainder) < EPSILON14 && Math.abs(angle) > EPSILON14) {
     return TWO_PI;
   }
   return remainder;
@@ -162,10 +179,10 @@ export function zeroToTwoPi(angle: number): number {
 /**
  * Produces an angle in the range -Pi <= angle <= Pi which is equivalent to the provided angle.
  *
- * @param {number} angle in radians
- * @returns {number} The angle in the range [<code>-PI</code>, <code>PI</code>].
+ * @param angle in radians
+ * @returns The angle in the range [<code>-PI</code>, <code>PI</code>].
  */
-export function negativePiToPi(angle: number): number {
+function negativePiToPi(angle: number): number {
   if (angle >= -PI && angle <= PI) {
     return angle;
   }
