+++ modules/culling/package.json
@@ -43,8 +43,7 @@
   ],
   "dependencies": {
     "@math.gl/core": "4.1.0-alpha.9",
-    "@math.gl/types": "4.1.0-alpha.9",
-    "@math.gl/geospatial": "4.1.0-alpha.9"
+    "@math.gl/types": "4.1.0-alpha.9"
   },
   "gitHead": "e1a95300cb225a90da6e90333d4adf290f7ba501"
 }
