+++ modules/culling/src/index.ts
@@ -10,7 +10,6 @@ export {OrientedBoundingBox} from './lib/bounding-volumes/oriented-bounding-box'
 export {CullingVolume} from './lib/culling-volume';
 export {Plane} from './lib/plane';
 export {Ray} from './lib/ray';
-export {EllipsoidTangentPlane} from './lib/ellipsoid-tangent-plane';
 
 export {PerspectiveOffCenterFrustum as _PerspectiveOffCenterFrustum} from './lib/perspective-off-center-frustum';
 export {PerspectiveFrustum as _PerspectiveFrustum} from './lib/perspective-frustum';
