+++ modules/geospatial/src/index.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-export {Ellipsoid} from './ellipsoid/ellipsoid';
-export {Rectangle} from './rectangle';
+export {Ellipsoid} from './ellipsoid';
+export {EllipsoidTangentPlane} from './ellipsoid-tangent-plane';
+export {LngLatRectangle} from './lng-lat-rectangle';
 export {isWGS84} from './type-utils';
